requirements.txt
# Core dependencies
apache-airflow==2.7.0
apache-airflow-providers-amazon==8.5.0
apache-airflow-providers-snowflake==5.0.0
requests==2.31.0
pandas==2.1.0
python-dotenv==1.0.0

# AWS
boto3==1.28.0
botocore==1.31.0

# Snowflake
snowflake-connector-python==3.2.0
snowflake-sqlalchemy==1.5.0

# Database
psycopg2-binary==2.9.7

# Utilities
pytz==2023.3